Make sequence identifiers into seperate package
