+++ docs/pydoc-markdown.yml
@@ -1,6 +1,6 @@
 loaders:
   - type: python
-    modules: [fasta, fastq, blast6, sequences]
+    modules: [fasta, fastq, blast6, sequences, sequence_identifiers]
     search_path: [../hypothesis_bio]
 processors:
   - type: pydocmd
